Separated FeedbackController out of PlayerController to support rumble during game over state
