+++ Assets/Samples/RebindingUI/Game/FeedbackController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8ac9b43eedbd9426f94712efe5efc1ca
